Add EBook class and update PaperBook constructor to use double for price

+++ Quantum Bookstore/Models/PaperBook.cs
@@ -4,7 +4,7 @@ public class PaperBook : Book
 {
     public int Stock {get; set;}
     
-    public PaperBook(string title, string author, string publisher, int year, int price, int stock)  
+    public PaperBook(string title, string author, string publisher, int year, double price, int stock)  
         : base(title, author, publisher, year, price)
     {
         if(stock<0) throw new Exception("Stock cannot be negative");
